+++ HISTORY.rst
@@ -19,6 +19,8 @@ PlatformIO 2.0
 * New `Microchip PIC32 <http://docs.platformio.org/en/latest/platforms/microchippic32.html>`__
   development platform with support for 20+ different PIC32 based boards
   (`issue #438 <https://github.com/platformio/platformio/issues/438>`_)
+* New RTOS and build Framework named `Simba <http://docs.platformio.org/en/latest/frameworks/simba.html>`__
+  (`issue #412 <https://github.com/platformio/platformio/issues/412>`_)
 * New boards for `ARM mbed <http://docs.platformio.org/en/latest/frameworks/mbed.html>`__
   framework: ST Nucleo F410RB, ST Nucleo L073RZ and BBC micro:bit
 * Added support for Generic ATTiny boards: ATTiny24, ATTiny25, ATTiny45 and